+++ src/main/java/com/sesac/carematching/caregiver/Status.java
@@ -1,0 +1,5 @@
+package com.sesac.carematching.caregiver;
+
+public enum Status {
+    OPEN, CLOSE
+}
