+++ src/main/java/com/sesac/carematching/caregiver/Type.java
@@ -1,0 +1,5 @@
+package com.sesac.carematching.caregiver;
+
+public enum Type {
+    HIRING, SEEKING
+}
