+++ src/main/java/com/sesac/carematching/caregiver/WorkForm.java
@@ -1,0 +1,5 @@
+package com.sesac.carematching.caregiver;
+
+public enum WorkForm {
+    COMMUTE, LIVE_IN
+}
