--- src/main/java/com/sesac/carematching/carematch/request/Status.java
@@ -1,5 +1,0 @@
-package com.sesac.carematching.carematch.request;
-
-public enum Status {
-    PENDING, APPROVED, REJECTED
-}
