--- src/main/java/com/sesac/carematching/carematch/request/Type.java
@@ -1,5 +1,0 @@
-package com.sesac.carematching.carematch.request;
-
-public enum Type {
-    HIRING, SEEKING
-}
