--- src/main/java/com/sesac/carematching/carematch/request/WorkForm.java
@@ -1,5 +1,0 @@
-package com.sesac.carematching.carematch.request;
-
-public enum WorkForm {
-    COMMUTE, LIVE_IN
-}
